Added weapons and skills

+++ Services/CharacterService/CharacterService.cs
@@ -1,4 +1,5 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
 using dotnet_rpg.Data;
 using dotnet_rpg.Dtos.Character;
 using Microsoft.EntityFrameworkCore;
@@ -9,18 +10,25 @@ public class CharacterService : ICharacterService
 {
     private readonly IMapper _mapper;
     private readonly DataContext _context;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public CharacterService(IMapper mapper, DataContext context)
+    public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
     {
         _mapper = mapper;
         _context = context;
+        _httpContextAccessor = httpContextAccessor;
     }
+
+    private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User
+        .FindFirstValue(ClaimTypes.NameIdentifier));
     
-    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
     {
         var response = new ServiceResponse<List<GetCharacterDto>>();
         var dbCharacters = await _context.Characters
-            .Where(c => c.User.Id == userId)
+            .Include(c => c.Weapon)
+            .Include(c => c.Skills)
+            .Where(c => c.User.Id == GetUserId())
             .ToListAsync();
         response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
         return response;
@@ -29,7 +37,10 @@ public class CharacterService : ICharacterService
     public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
     {
         var serviceResponse = new ServiceResponse<GetCharacterDto>();
-        var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+        var dbCharacter = await _context.Characters
+            .Include(c => c.Weapon)
+            .Include(c => c.Skills)
+            .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
         serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
         return serviceResponse;
     }
@@ -38,9 +49,13 @@ public class CharacterService : ICharacterService
     {
         var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
         Character character = _mapper.Map<Character>(newCharacter);
+        character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+        
         _context.Characters.Add(character);
         await _context.SaveChangesAsync();
-        serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+        serviceResponse.Data = await _context.Characters
+            .Where(c => c.User.Id == GetUserId())
+            .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
         return serviceResponse;
     }
 
@@ -50,20 +65,29 @@ public class CharacterService : ICharacterService
 
         try
         {
-            var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-
-            _mapper.Map(updatedCharacter, character);
-            // character.Name = updatedCharacter.Name;
-            // character.HitPoints = updatedCharacter.HitPoints;
-            // character.Strength = updatedCharacter.Strength;
-            // character.Defense = updatedCharacter.Defense;
-            // character.Intelligence = updatedCharacter.Intelligence;
-            // character.Class = updatedCharacter.Class;
-
-            await _context.SaveChangesAsync();
-
-            response.Data = _mapper.Map<GetCharacterDto>(character);
+            var character = await _context.Characters
+                .Include(c => c.User)
+                .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
             
+            if (character.User.Id == GetUserId())
+            {
+                _mapper.Map(updatedCharacter, character);
+                // character.Name = updatedCharacter.Name;
+                // character.HitPoints = updatedCharacter.HitPoints;
+                // character.Strength = updatedCharacter.Strength;
+                // character.Defense = updatedCharacter.Defense;
+                // character.Intelligence = updatedCharacter.Intelligence;
+                // character.Class = updatedCharacter.Class;
+
+                await _context.SaveChangesAsync();
+
+                response.Data = _mapper.Map<GetCharacterDto>(character);
+            }
+            else
+            {
+                response.Success = false;
+                response.Message = "Character not found.";
+            }
         }
         catch (Exception ex)
         {
@@ -80,16 +104,68 @@ public class CharacterService : ICharacterService
 
         try
         {
-            Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-            _context.Characters.Remove(character);
-            await _context.SaveChangesAsync();
-            response.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            Character character = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+            if (character != null)
+            {
+                _context.Characters.Remove(character);
+                await _context.SaveChangesAsync();
+                response.Data = _context.Characters
+                    .Where(c => c.User.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            }
+            else
+            {
+                response.Success = false;
+                response.Message = "Character not found.";
+            }
+            
         }
         catch (Exception ex)
         {
             response.Success = false;
             response.Message = ex.Message;
         }
+        return response;
+    }
+
+    public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto newCharacterSkill)
+    {
+        var response = new ServiceResponse<GetCharacterDto>();
+        try
+        {
+            var character = await _context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .FirstOrDefaultAsync(c =>c.Id == newCharacterSkill.CharacterId &&
+                                         c.User.Id == GetUserId());
+
+
+            if (character == null)
+            {
+                response.Success = false;
+                response.Message = "character not found.";
+                return response;
+            }
+
+            var skill = await _context.Skills.FirstOrDefaultAsync(s => s.Id == newCharacterSkill.SkillId);
+            if (skill == null)
+            {
+                response.Success = false;
+                response.Message = "skill not found.";
+                return response;
+            }
+            
+            character.Skills.Add(skill);
+            await _context.SaveChangesAsync();
+            response.Data = _mapper.Map<GetCharacterDto>(character);
+        }
+        catch (Exception e)
+        {
+            response.Success = false;
+            response.Message = e.Message;
+        }
+
         return response;
     }
 }
